28/03/2026's commit

+++ app/Http/Controllers/hr_manager/EmployeesController.php
@@ -312,7 +312,9 @@ class EmployeesController extends Controller
                         unlink($imagePath); // Delete the image
                     }
                 }
-                $user->userInfo->delete();
+                if ($user->userInfo) {
+                    $user->userInfo->delete();
+                }
                 $user->delete();
             });
             flash()->options([
